use fixture for web vitals to encourage reuse
